+++ src/effects/Reverse.cpp
@@ -114,7 +114,7 @@ bool EffectReverse::ProcessOneWave(int count, WaveTrack * track, sampleCount sta
    // perform a split at the start and/or end of the reverse selection
    const auto &clips = track->GetClips();
    // Beware, the array grows as we loop over it.  Use integer subscripts, not iterators.
-   for (int ii = 0; ii < clips.size(); ++ii) {
+   for (size_t ii = 0; ii < clips.size(); ++ii) {
       const auto &clip = clips[ii].get();
       auto clipStart = clip->GetStartSample();
       auto clipEnd = clip->GetEndSample();
