+++ src/effects/ladspa/LadspaEffect.cpp
@@ -906,12 +906,12 @@ bool LadspaEffect::ProcessFinalize()
 
 size_t LadspaEffect::ProcessBlock(float **inBlock, float **outBlock, size_t blockLen)
 {
-   for (int i = 0; i < mAudioIns; i++)
+   for (int i = 0; i < (int)mAudioIns; i++)
    {
       mData->connect_port(mMaster, mInputPorts[i], inBlock[i]);
    }
 
-   for (int i = 0; i < mAudioOuts; i++)
+   for (int i = 0; i < (int)mAudioOuts; i++)
    {
       mData->connect_port(mMaster, mOutputPorts[i], outBlock[i]);
    }
@@ -972,12 +972,12 @@ size_t LadspaEffect::RealtimeProcess(int group,
                                           float **outbuf,
                                           size_t numSamples)
 {
-   for (int i = 0; i < mAudioIns; i++)
+   for (int i = 0; i < (int)mAudioIns; i++)
    {
       mData->connect_port(mSlaves[group], mInputPorts[i], inbuf[i]);
    }
 
-   for (int i = 0; i < mAudioOuts; i++)
+   for (int i = 0; i < (int)mAudioOuts; i++)
    {
       mData->connect_port(mSlaves[group], mOutputPorts[i], outbuf[i]);
    }
