+++ src/export/ExportMP3.cpp
@@ -1855,7 +1855,7 @@ ProgressResult ExportMP3::Export(AudacityProject *project,
 
          short *mixed = (short *)mixer->GetBuffer();
 
-         if (blockLen < inSamples) {
+         if (blockLen < (unsigned int)inSamples) {
             if (channels > 1) {
                bytes = exporter.EncodeRemainder(mixed, blockLen, buffer.get());
             }
